Grove_SI114X.CPP
Hello,
So I'm quite new to Wio link for a week now and I'm working hard to make my sunlight and dust sensor working.
So here is just a copy of the arduino grove driver of sunlight sensor with the right header I supposed for Wio_link. I'm not expert in that but tried to do the right way.
I wished to test with my own server but still in troubles to launch it. (I have separate topics here : )
Thanks,
Peter